Remove misleading `Eq` instance
The `Eq` instance on Priority does not work as one would initially expect: for example, two `PrioLowest` are considered unequal.
I'm not sure whether this behaviour is required. As the derived Eq instance on `Production` also features this oddity and is not actually used anywhere in happy, we remove it for the better.